+++ src/main/java/auth/config/LoginMember.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.config;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
