+++ src/main/java/auth/config/MemberDetails.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.config;
 
 public interface MemberDetails {
     Long getId();
