+++ src/main/java/auth/dto/TokenResponse.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.dto;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
