+++ src/main/java/nextstep/member/MemberController.java
@@ -1,6 +1,6 @@
 package nextstep.member;
 
-import auth.LoginMember;
+import auth.config.LoginMember;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
