+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,6 +1,6 @@
 package nextstep.reservation;
 
-import auth.AuthenticationException;
+import auth.exception.AuthenticationException;
 import nextstep.member.Member;
 import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
