+++ src/main/java/nextstep/support/BaseExceptionHandler.java
@@ -1,6 +1,6 @@
 package nextstep.support;
 
-import auth.AuthenticationException;
+import auth.exception.AuthenticationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
